Base User, Devise and Page Setup
Installed the Devise gem, edited several Devise related files to enhance
Devise's usability in the application.

Created a Devise user model with the standard Devise generated configuration.

Generated Devise views that will feature for anything user login, sign up etc .
related.

Added a root route as pages#home and created the view with just a title
for now to display on the homepage. Also created a pages controller to
allow the home specific page to display.