Post Model Setup and DB Updated
Created a post model that I envision to be the basis
for forum posts, including the original post and subsequent
replies to it.

Posts belong to users, however provided an optional association for
posts to belong to a team, as in if a user wants to specify that the post
relates to a specific team.

Included a validation for content, so that users creating a forum post
must be entering text in the content field.

A likes counter has been implemented in the post model too, with a method
to define how likes are count, a future likes model will implement this.

Likely to also add further validations and associations in this post model
going forward.