Like Model Setup, Like Controller Created
Setup a like model for handling likes in the application, along with a
DB migration into the schema.

Associations and conditions setup in the like model to ensure likes are counted
per like record, associated with posts and users along with being able to only
provide one like per user for a post.

Also generated the likes controller which will be built upon later on.