Noticed Gem Setup along with User Mailer
Setup latest version (2.4.0) of the noticed gem to
handle post notifications when posts are created.

A user mailer has also been created to handle the greeting
for the post notification and who it is emailed to.

Also params have been setup to determine that the user will be
the recipient of the new post notifications.