Discussions Setup and Links Added
Added MVC for discussions, posts belong to a discussion
and a discussion can be seen as a thread.

Created a migration file to enhance the association between discussions
and posts, therefore a discussion ID is now associated with a post.

The discussions index base view page version appears along
with the simplest possible empty index method in the discussions
controller in order to be able to render the discussions index
page.

Links added to discussions index and Devise sign in page,
with an if statement also included in the same view page
for if the user is signed in or not will affect how the
page appears for them.

The routes have also been updated to include CRUD resources for discussions
and nested resources of posts inside discussions.