Setup Bootstrap, Navbar, Footer and Other Pages
Bootstrap and sassc gems latest versions added to the
Gemfile and installed into the application. Have also
added Bootstrap into many key files to setup it to
appear, and also renamed the application.css to
application.scss to handle assets/stylesheets going
forward.

A new footer partial scss file has been added, for now
designed to keep footers at the bottom of the page.

A navbar and footer have partial views created for
them and appear across the app, some of the navbar
has been setup but the footer has just placeholder
text for now.

Reduced the size of the home along with updating it
to not duplicate the app name (which now appears in
the navbar) and moved the Devise user sign in/out etc.
to the navbar, and also added the ability to edit the
user profile for the specific user who has logged in.

Reconfigured the Devise initializer file to allow
the sign_out link when pressed to sign out and not
run into an error.

Also added code for the index method for the discussions
controller so the index page will display all discussions.