Incorporated "Current" to handle Current User
To take advantage of Devise's current user method, a
new "current" model has been created with a user attribute
to allow current user to be called elsewhere in the application
to access the current user. It has also been added to the
application controller and discussion model, and has been
verified via the console.

Within the discussions controller current user has been called in the
new create method (along with other parts including a discussion save and
a setup for the discussion creation being dependent upon the discussion
params) and an empty new method for handling if the discussion create process
is not succesful for whichever reason for the user.

Also added a new current suer email method to the navbar and removed the
welcome message so that the edit user path only shows instance of the
current user's email address and which when clicked allows the user to
edit their user account.