New Discussion Setup, Simple Form Gem Added
Revised the navbar to introduce the new discussion path
rather than the root path, so clicking the link goes to
the new discussion page. In the same page added in
the view code for rendering discussions in an if statement,
while the else displays a notice message when no discussions
have been rendered.

Added the latest version of the Simple Form gem (5.3.1)
and installed the Simple Form Bootstrap version for parity
with Bootstrap in the application.

A new view ERB page has been setup with a title and ability to render a form,
which also means a new form partial has been created, which will incorporate
Simple Form (as introduced above).

Also added new code for the new discussion controller method.