Discussion Now able to be Created
Within a Simple Form structure, logged in users are
now able to create a discussion just with entering a
title and submitting it at this stage.

Altered the index page to have a button correctly
appear.

Added to the discussions controller and discussion
model to handle title validation and error messages
if a title is not entered when creating a discussion.