Edit Discussions Setup
Implemented the ability to edit an existing discussion.

Added edit and update discussions_controller actions
to handle the edit and update occurring, while a private
controller method also sets the discussion and as per the new
before_action is only called when editing and updating a discussion.

A link to edit a discussion exists now within the discussion partial also.

New edit view introduced too for rendering the edit function in the
frontend of the application.