Delete Discussions Implemented
The first Turbo implementation in the application, to
delete a discussion has been created.

When a discussion link_to delete is pressed, it deletes via
turbo intercepting the click and issuing a DELETE request to the server.

A delete method in the controller has been setup for handling delete discussion
requests, along with a basic discussion show page for displaying active
discussions.