Show Discussions Setup With Turbo also Enhancing the App
Created a new partial for handling the discussion content being shown,
therefore in a show, edit or delete view for a dicussion the user will
be able to view the discussion during each action.

Updated the edit and update actions in the discussions controller to
handle redirects better.

The edit page has been updated to allow to cancel via a link which redirects
the user to the specific discussion.

Several Turbo Frame Tags now encapsulate the new show and existing edit
pages for discussions.