Enhanced functionality of post and discussion models
Posts can now be created, associated with a specific
discussion by the current user, and validations check
for the presence of characters and do not allow empty
posts to be created.

A posts count has also been introduced to the application,
sharing similarities with the likes count that was previously
introduced in the posts model.