Post Added to New Discussions
Users are now required to add an initial post to the
discussion when creating it.

The post nested within a new discussion has been added
to the discussion model and rendered inside a Trix
rich text editor box in the discussion form partial.

Also fixed an error with how the Trix was being imported.