Posts Show and Layout Adjustments Made
Posts are now visible in the discussion page they
were created in via a new page partial.

Posts are timestamped as to when they were made in the newly created
posts header partial that is called in the discussions show page and
subsequently displays there.

The discussions index have had the top of the layout altered, a button
now contains the newly revised worded "Create a Discussion" button link,
and it appears opposite the Active Discussions title on the page.

Also added a Back to Discussions button to allow the user
to return to the discussions index.