Posts Now Display in Realtime Once Created
Posts are now appearing to the current broswer tab
and other browser tabs in realtime.

A few layout changes have been made, enhancing the application's
structure ahead of further changes going forward.

Turbo has been introduced in several more places to help with the realtime
rendering of posts created.

The issue is new posts are appearing below the Create Post and Back to
Discussions buttons at the bottom of the page, will look to resolve this.