Posts now Edited and Created Update in all places in realtime
Incorporated the edit post function, that appears in the same show discussion
and post page as posts that are shown and also created. Once edited and
edit confirmed by the user they are rendered in their updated format in
realtime to the current user and other users on the same page.

Also fixed the show page error where new posts would appear below the
create post form rather than the intended position of above the form with
already existing posts, they now appear in the correct place with existing
posts in the specific discussion.