Posts are now Deleted in Real Time
A prompt now appears when a delete link is clicked by
the user, once confirmed on the notice prompt the post
is then deleted in realtime to the user and others
without a page refresh being necessary.

A new Turbo frame is inserted in the deletion link data
along with a Turbo confirmation message.