Categories Added to the Forum
A generated scaffold for Category has allowed a lot of the code
in this application to prebuilt by the scaffold command.

The navbar has been updated with links to the Discussions and Categories
indexes.

Categories has been associated with the discussion model, and when creating a
new discussion it needs to be assigned a category, that the user can also
create (categories) via the category index page.

Migrations have also been performed and integrated into the DB schema to
bring in categories and a category ID being assigned to a specific discussion.

The discussion form has also added the new category collection select to it.