Categories Added to the Discussion Index
Categories are now visible along with a count of the
amount of discussions in each category in the dicussions
index view page.

Turbo broadcasts have been implemented in the category model
to display realtime changes to the user and others without refresh,
on the discussion index page.

There is also a posts count in the discussion index displaying the
amount of posts per discussion, a good way for users to see if there
have been any new posts to a discussions without having to view the
page to check.