Category Filtering Added to discussions
Categories can now be filtered to just the discussions that
they contain, and discussions created are added to their specific
category chosen.

Categories can be selected from the discussions index and then display
only the discussions created for that particular category.

The routes have also been updated to assign category ID's to the discussions
index.