Discussions can be Closed to Prevent Further Posts
Discussions can be closed when created and via the edit
action to prevent additional posts from being created within them.

Added the closed column to the discussions table in the schema via a
new migration.

The discussion form has been updated to allow the user to select whether
or not the discussion is closed or not, this is not permanent and can be
ticked and unticked at the user's desire.