Added Lazy Loading to the Posts Partial
Using Turbo means it automatically fetches content and in the case of Turbo
Frames, automatically fetches and replaces the frame too. With lazy loading
implemented, posts are not automatically fetched with all of the post actions:
Edit and Delete. Rather, they are only loaded when they are needed, which
results in a Turbo Frame with a source URL being created. This frame will
load its content from discussion_post_path(post.discussion, post) when it
becomes visible or is otherwise triggered.

New admin column added via a migration to the user schema to allow the
admin user to be usable.