Noticed Notifier Setup
Several models have been created to handle the new implementation
of notifications into the app.

The previously added Noticed gem is the basis of the implementation,
with new post notifier and notification files woring as the backend to
allow notifications to be delivered to users, with code therefore also
introduced in the user and post models for when to notify users of a new
post.