Email Notifications Now Send on New Posts in Discussions
If a user is subscribed to a discussion and a new post is
made in the post by another user, they are notified now also by
email.

Also I had an issue relating to name instead of title in a line of code in the
previous PR related to notifications, this has now been resolved and notifications
appear in the notifications index page for a user dependent upon the activity in a
discussion they are subscribed to, and links directly to the discussion where the
new post has been made.