Authentication Enhanced before being able to Navigate
Strengthened the authentication around what a non
authenticated user can do before logging in or first
signing up.  This is to prevent a user from being able to
navigate to the discussions page without being logged in.  This
is a security enhancement.  The user will be redirected to the
login page if they try to access the discussions page without
being logged in.

In the categories controller I added the authenticate_user!
method to the before_action. This is to ensure that only
authenticated users can access the categories page.